add validation, fix tests and comment resolution

+++ mainflux/bootstrap.js
@@ -8,9 +8,6 @@ class  Bootstrap{
      * It is used to create, update, view and remove bootstrap configurations.
      * It is also used to bootstrap a thing.
      * @param {string} bootstraps_url - The url of the bootstraps service.
-     * @param {string} content_type - The content type of the request.
-     * @param {string} bootstrapsEndpoint - The endpoint of the bootstraps service which is 
-     * configs.
      * @returns {Bootstrap} - Returns a Bootstrap object.
      * 
      */
@@ -18,6 +15,18 @@ class  Bootstrap{
         this.bootstraps_url = bootstraps_url;
         this.content_type = "application/json";
         this.bootstrapsEndpoint = "configs";
+    }
+
+    ValidateConfigAndToken(config, token){
+    //Validate config
+        if (typeof config !== "object" || config === null) {
+            throw new Error('Invalid config parameter. Expected an object.');
+        }
+
+        // Validate token
+        if (typeof token !== "string" || token === null) {
+            throw new Error('Invalid token parameter. Expected a string.');
+        }    
     }
 
     Create(config, token){
@@ -37,9 +46,12 @@ class  Bootstrap{
          *      "name": "thing_name"
          *   } 
          */
+
+        this.ValidateConfigAndToken(config, token);
+
         const options = {
             method: "post",
-            maxBodyLength: Infinity,
+            maxBodyLength: 2000,
             url: `${this.bootstraps_url}/things/${this.bootstrapsEndpoint}`,
             headers: {
                 "Content-Type": this.content_type,
@@ -71,9 +83,12 @@ class  Bootstrap{
          *      "name": "thing_name"
          * }
          */
+
+        this.ValidateConfigAndToken(config, token);
+
         const options = {
             method: "put",
-            maxBodyLength: Infinity,
+            maxBodyLength: 2000,
             url: `${this.bootstraps_url}/things/state/${config["thing_id"]}`,
             headers: {
                 "Content-Type": this.content_type,
@@ -105,9 +120,12 @@ class  Bootstrap{
          *      "name": "thing_name"
          * }
          */
+
+        this.ValidateConfigAndToken(config, token);
+
         const options = {
             method: "put",
-            maxBodyLength: Infinity,
+            maxBodyLength: 2000,
             url: `${this.bootstraps_url}/things/configs/${config["thing_id"]}`,
             headers: {
                 "Content-Type": this.content_type,
@@ -132,9 +150,16 @@ class  Bootstrap{
          * @param {string} thing_id - The thing_id of the configuration to be viewed.
          * @param {string} token - The token to be used for authentication.
          */
+
+        if (typeof thing_id !== "string" || thing_id === null) {
+            throw new Error('Invalid thing_id parameter. Expected a string.');
+        }
+
+        this.ValidateConfigAndToken({}, token);
+
         const options = {
             method: "get",
-            maxBodyLength: Infinity,
+            maxBodyLength: 2000,
             url: `${this.bootstraps_url}/things/${this.bootstrapsEndpoint}/${thing_id}`,
             headers: {
                 "Content-Type": this.content_type,
@@ -162,6 +187,15 @@ class  Bootstrap{
          * @param {string} token - The token to be used for authentication.
          * 
          */
+
+        if (typeof config_id !== "string" ||
+            typeof client_cert !== "string" ||
+            typeof client_key !== "string" ||
+            typeof ca !== "string" ||
+            typeof token !== "string" ) {
+            throw new Error('Invalid parameter types. Expected strings for config_id, client_cert, client_key, ca and token.');
+        };
+
         const payload = {
             "client_cert": client_cert,
             "client_key": client_key,
@@ -169,7 +203,7 @@ class  Bootstrap{
         }
         const options = {
             method: "patch",
-            maxBodyLength: Infinity,
+            maxBodyLength: 2000,
             url: `${this.bootstraps_url}/configs/certs/${config_id}`,
             headers: {
                 "Content-Type": this.content_type,
@@ -195,9 +229,16 @@ class  Bootstrap{
          * @param {string} token - The token to be used for authentication.
          * 
          */
+
+        if (typeof config_id !== "string" || config_id === null) {
+            throw new Error('Invalid config_id parameter. Expected a string.');
+        }
+
+        this.ValidateConfigAndToken({}, token);
+
         const options = {
             method: "delete",
-            maxBodyLength: Infinity,
+            maxBodyLength: 2000,
             url: `${this.bootstraps_url}/things/${this.bootstrapsEndpoint}/${config_id}`,
             headers: {
                 "Content-Type": this.content_type,
@@ -221,9 +262,14 @@ class  Bootstrap{
          * @param {string} external_key - The encrypted external key of the configuration to be retrieved.
          * @return {object} - Returns a config object.
          */
+
+        if (typeof external_id !== "string" || typeof external_key !== "string") {
+            throw new Error('Invalid type of parameters. Expected strings for external_key and external_id.');
+        }
+        
         const options = {
             method: "get",
-            maxBodyLength: Infinity,
+            maxBodyLength: 2000,
             url: `${this.bootstraps_url}/things/bootstrap/${external_id}`,
             headers: {
                 "Content-Type": this.content_type,
